requirements-hs.txt
comfyui-frontend-package>=1.23.4
comfyui-workflow-templates>=0.1.47
comfyui-embedded-docs>=0.2.4
torch
torchvision
torchdiffeq>=0.2.3
torchsde>=0.2.6
einops>=0.6.0
open-clip-torch>=2.24.0
transformers!=4.53.0,!=4.53.1,!=4.53.2
tokenizers>=0.13.3
sentencepiece
peft>=0.10.0
torchinfo
safetensors>=0.4.2
aiohttp>=3.11.8
yarl>=1.9.4
accelerate>=0.25.0
pyyaml>=6.0
scikit-image>=0.20.0
jsonmerge>=1.9.0
clean-fid>=0.1.35
resize-right>=0.0.2
opencv-python-headless>=4.9.0.80,<4.10.0
albumentations>=1.3.0
aiofiles>=23.1.0
frozendict>=2.3.6
python-dateutil>=2.8.2
importlib_resources
Pillow
scipy
tqdm
protobuf>=3.20.0,<5.0.0
psutil
ConfigArgParse
aio-pika
pyjwt[crypto]
kornia>=0.7.0
mpmath>=1.0,!=1.4.0a0
huggingface_hub[hf_xet]>=0.32.0
lazy-object-proxy
lazy_loader>=0.3
can_ada
fsspec
natsort
OpenEXR
opentelemetry-distro
opentelemetry-sdk<1.34.0
opentelemetry-exporter-otlp<=1.27.0
opentelemetry-propagator-jaeger
opentelemetry-instrumentation
opentelemetry-util-http
opentelemetry-instrumentation-aio-pika
opentelemetry-instrumentation-requests
opentelemetry-semantic-conventions
wrapt>=1.16.0
certifi
spandrel>=0.3.4
numpy>=1.24.4,<2.3  # See https://github.com/conda-forge/numba-feedstock/issues/158
soundfile
watchdog
PySoundFile
networkx>=2.6.3
joblib
jaxtyping
spandrel_extra_arches
ml_dtypes
diffusers>=0.30.1
vtracer
skia-python
pebble>=5.0.7
openai
anthropic
humanize
lightning
flax
jax
colour
av>=14.2.0
pydantic~=2.0
pydantic-settings~=2.0
typer
ijson
scikit-learn>=1.4.1
setuptools
alembic
SQLAlchemy
gguf